requirements.txt
streamlit>=1.35.0
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0
plotly>=5.18.0
scikit-learn>=1.3.0
matplotlib>=3.7.0


